add methods and test for inserting/querying notes with labels

+++ app/src/main/java/com/github/androidpirate/slicknotes/data/NoteWithLabels.java
@@ -43,7 +43,7 @@ public class NoteWithLabels {
         return note;
     }
 
-    public int getNoteId() {
+    public long getNoteId() {
         return note.getNoteId();
     }
